requirements.txt
google-auth-oauthlib>=0.5.0
google-api-python-client>=2.0.0
requests>=2.28.0
python-dotenv>=0.21.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
email-validator>=2.0.0
schedule>=1.2.0
loguru>=0.7.0